requirements.txt
zmq
urllib3
requests
pycryptodome
